Bug fix & New menu option: 'ICONV Ignore'
Fixed problem when loading a project file that uses '\n' together with
'\r' in empty texts. '\r' was causing the program to create extra lines.
Now program ignores them and all the '\n' if present at the beginning of
a text.
Also 'ICONV Ignore' has been added, to be used when some hidden special
characters appear to be in some texts. They caused program to crash when
ICONV was not in use (not fixed).